Aumentado a versão minima do PHP na LIB. Incluido variavel de ambiente do PHPStan Criado uma pasta data, para armazenar casos maiores de massa de dados de teste Retirado mixed que só funciona na versão 8.0 do PHP ou superior Adicionado pasta tests na analise do PHPStan
