+++ src/DependencyInjection/FormatAux.php
@@ -14,7 +14,7 @@ abstract class FormatAux
     /**
      * @param mixed $val
      */
-    private static function returnTypeBool(mixed $val, bool $returnNull = false): bool
+    private static function returnTypeBool($val, bool $returnNull = false): bool
     {
         $boolVal = (is_string($val) ?
             filter_var($val, FILTER_VALIDATE_BOOLEAN, FILTER_NULL_ON_FAILURE) : (bool) $val);
